Completly new structur
The second version had all going and made it difficult to show rainbow colors with full controll of the button and the rotary encoder. I also tried pseudo "Threads", but they doesn't work and were hell of a mess. Here I tried to use a structure to use one Array that always get updated to the LEDs after a loop ended. It's also easier to understand what the system is doing. So if I want to be in rainbow mod I can update the Arrays for the LEDs for the next round let the controller set the LEDs and still have the controll of Button and Rotary Encoder.